requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
streamlit>=1.28.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-classic>=0.1.0
langchain-google-genai>=0.0.1
langchain-huggingface>=0.0.1
langchain-community>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.0.0
langgraph>=0.1.0
google-generativeai>=0.3.0
pypdf>=3.17.0
pydantic>=2.0.0
python-dotenv>=1.0.0
